finished the scene loading api for c#; fixed a bunch of warnings

+++ TerranEditor/src/EditorPanels/LogPanel.cpp
@@ -62,7 +62,7 @@ namespace TerranEditor
 			UI::ScopedFont mediumFont("Roboto-Regular_Medium");
 			
 			ImGuiListClipper logTextClipper;
-			logTextClipper.Begin(m_TextBuffer.size());
+			logTextClipper.Begin(static_cast<int>(m_TextBuffer.size()));
 
 			while(logTextClipper.Step()) 
 			{
